ppcstats: Move a bunch of structures to separate file
Reduces clutter.
